Initial Source Upload of 2.6.32

+++ arch/ia64/kernel/efi.c
@@ -23,11 +23,9 @@
  */
 #include <linux/module.h>
 #include <linux/bootmem.h>
-#include <linux/crash_dump.h>
 #include <linux/kernel.h>
 #include <linux/init.h>
 #include <linux/types.h>
-#include <linux/slab.h>
 #include <linux/time.h>
 #include <linux/efi.h>
 #include <linux/kexec.h>
@@ -39,7 +37,6 @@
 #include <asm/pgtable.h>
 #include <asm/processor.h>
 #include <asm/mca.h>
-#include <asm/setup.h>
 #include <asm/tlbflush.h>
 
 #define EFI_DEBUG	0
@@ -157,7 +154,7 @@ prefix##_get_next_variable (unsigned long *name_size, efi_char16_t *name,      \
 #define STUB_SET_VARIABLE(prefix, adjust_arg)				       \
 static efi_status_t							       \
 prefix##_set_variable (efi_char16_t *name, efi_guid_t *vendor,		       \
-		       u32 attr, unsigned long data_size,		       \
+		       unsigned long attr, unsigned long data_size,	       \
 		       void *data)					       \
 {									       \
 	struct ia64_fpreg fr[6];					       \
